feature/2: MainPage.tsx moved from components folder into pages
