+++ apps/local/src/app/app.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import MainPage from './components/MainPage';
+import MainPage from './pages/MainPage';
 
 export function App() {
   return (
